app/requirements.txt
uvicorn
fastapi
redis 
requests 
web3 
python-dotenv
numpy
# pymongo
# odmantic
pydantic
sqlmodel
psycopg2-binary
